requirements.txt
boto3
requests
pandas
jupyter
requests
python-dotenv